requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
google-auth==2.37.0
google-cloud-speech==2.30.0
google-cloud-aiplatform==1.74.0
httpx>=0.27.0
google-cloud-texttospeech>=2.16.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.100.0
aiosqlite>=0.19.0
PyPDF2>=3.0.0
